+++ NEWS.md
@@ -1,6 +1,14 @@
-# MVisAGe 0.1.0
+# MVisAGe 0.2.0
 
-* Added a `NEWS.md` file to track changes to the package.
+* The functions corr.compute() and corr.list.compute() now provide the option to calculate Spearman correlation coefficients.
+
+* The output of the functions corr.compute() and corr.list.compute() includes the t statistic, one-sided p-value, and Benjamini-Hochberg corrected p-value corresponding to the correlation coefficient.
+
+* The functions perm.significance() and perm.significance.list.compute() were added.
+
+* MVisAGe 0.2.0 was submitted to CRAN on August 22, 2017.
+
+# MVisAGe 0.1.0
 
 * MVisAGe 0.1.0 submitted to CRAN on May 1, 2017